refactor(scripts): split the two files that broke the 500-line rule they enforce
`check-doc-versions.mjs` (911 lines) and `test-guards.mjs` (943) both violated the
rule they exist to enforce, and this session added 128 lines to the first one. It
is not cosmetic: this session's defects clustered in the parts too large to hold
in one's head — a wrong failure signal for the bundle checker, a comment counted
as a mutation site, a `sinal` matching its own declaration.

Guard modules now mirror their test modules:

  check-doc-versions.mjs  387   guards/settings.mjs        288 <-> tests-settings.mjs
  test-guards.mjs         398   guards/versions.mjs        188
  test-harness.mjs        325   guards/derived-counts.mjs  148 <-> tests-derived-counts.mjs

No file exceeds 500 lines; the largest is 398.

The bodies moved verbatim, original indentation included. That is deliberate: the
file mixes template literals with backticks inside regex character classes
(`/[;|&`\n]/`), so no automated re-indentation is safe, and a verbatim move is
auditable against the strongest test a pure refactor has — byte-identical output.

Six criteria, all met: no file over 500 lines; 178 tests with identical names and
zero failures; the guard's output byte-identical at 1495 bytes; the mutation site
total preserved at 64 (20+3+7+9+6+14+5) and fully covered; identical results from
a subdirectory; and SIGKILL mid-sweep leaving all 14 files byte-identical.

Three defects surfaced, none by reading:

- The dependency-version guards used `pkgRaw`, declared inside Guard 3. Splitting
  the functions broke that sharing and byte-identical output caught it with a
  ReferenceError. A module-level memo restores the single read. My dependency
  analysis had only looked for helper names, not every free variable.

- The `CHECKS` test patched text in the entry point with `.replace()`. `CHECKS`
  moved, the replace stopped matching, and it returned the file untouched — a test
  that would pass for the wrong reason. `BANNED` had the same exposure. Both now
  throw when the target text is absent.

- The synthetic sandbox builds files rather than copying, so it never carried the
  new module and the suite aborted loudly. It now copies the whole `guards/`
  directory, so extracting more modules will not break it again.

A guard extracted into `guards/*.mjs` MUST get its `PARES` entry in the mutation
sweep: the warnings move with it, and without the entry the sweep covers only the
entry point and reports 100% while measuring less. Recorded in the propagation
matrix, with the rule that the site total must match across a refactor.

Closes #13. Nothing was written to any context file or the CHANGELOG.
