+++ .agent/rules/sync-docs.md
@@ -35,7 +35,10 @@ Nao basta atualizar apenas os ficheiros de contexto (`.agent/context/`) — e ob
     **Verificador novo ou alterado?** Entao (a) tem o seu `test-*.mjs` com controlos negativos,
     (b) esta registado em `PARES` no `mutation-sweep.mjs` com o seu `sinal` de reprovacao, e
     (c) `node .agent/scripts/mutation-sweep.mjs` sai 0. A varredura reprova de proposito um
-    verificador sem suite — nao a silenciar, escrever a suite
+    verificador sem suite — nao a silenciar, escrever a suite.
+    **Guard extraido para `guards/*.mjs`?** A entrada em `PARES` e obrigatoria: os avisos
+    passam a viver no modulo, e sem ela a varredura cobre so o ficheiro de entrada e reporta
+    100% a mentir. A soma dos sitios antes e depois de um refactor tem de ser a MESMA
 15. [ ] `src/docs/CHANGELOG.md` — versao atual registada (`## [vX.Y.Z] - Descricao`), alinhada com `package.json`
 16. [ ] `src/docs/` restantes — manuais refletem UI/logica atual
 17. [ ] `.github/workflows/ci.yml` — CI pipeline reflete comandos e targets atuais
