+++ .agent/scripts/check-doc-versions.mjs
@@ -32,6 +32,9 @@
 import { readFileSync, readdirSync } from "fs";
 import { fileURLToPath } from "url";
 import { dirname, resolve, join } from "path";
+import { guardSettings } from "./guards/settings.mjs";
+import { guardChangelogVersion, guardDependencyVersions } from "./guards/versions.mjs";
+import { guardDerivedCounts } from "./guards/derived-counts.mjs";
 
 // --- Ancoragem a raiz do repo -------------------------------------------------
 // Este ficheiro vive em <raiz>/.agent/scripts/, logo a raiz esta dois niveis acima.
@@ -166,89 +169,8 @@ if (claude !== null && gemini !== null) {
 }
 
 // --- Guard 3: package.json version === ultima versao do CHANGELOG ---
-// So se aplica quando ja existe package.json. A partir dai, tudo o que falte e WARN:
-// um CHANGELOG movido ou um package.json sem `version` desligavam este guard em silencio.
-const CHANGELOG_PATH = "src/docs/CHANGELOG.md";
-const pkgRaw = read("package.json");
-if (pkgRaw === null) {
-  skip("Guard 3 (versoes) — sem package.json (template ainda sem app)");
-} else {
-  let pkgVersion = null;
-  let pkgInvalid = false;
-  try {
-    pkgVersion = JSON.parse(pkgRaw).version ?? null;
-  } catch {
-    pkgInvalid = true;
-    warn("package.json invalido — nao foi possivel ler version");
-  }
-  const changelog = read(CHANGELOG_PATH);
-  if (!pkgInvalid && !pkgVersion) {
-    warn(`package.json sem campo "version" utilizavel — o Guard 3 nao tem com que comparar`);
-  }
-  if (changelog === null) {
-    warn(`${CHANGELOG_PATH} nao encontrado — core-rules.md exige-o atualizado antes de cada commit`);
-  } else if (pkgVersion) {
-    // Ignorar exemplos dentro de comentarios HTML (ex: o `## [v0.1.0]` de exemplo no template).
-    const clNoComments = changelog.replace(/<!--[\s\S]*?-->/g, "");
-    // Aceita `## [v1.2.3]`, `## v1.2.3` e `## 1.2.3`, com sufixos SemVer (pre-release/build).
-    const SEMVER = /^##\s*\[?v?([0-9]+\.[0-9]+\.[0-9]+(?:[-+][0-9A-Za-z.-]+)?)\]?/gm;
-    const found = [...clNoComments.matchAll(SEMVER)].map((m) => m[1]);
-    const headings = [...clNoComments.matchAll(/^##\s+\S/gm)].length;
-    if (found.length === 0) {
-      if (headings > 0) {
-        warn(`${CHANGELOG_PATH} tem ${headings} heading(s) "##" mas nenhum no formato "## [vX.Y.Z] - Descricao" — formato invalido`);
-      } else {
-        note(`${CHANGELOG_PATH} ainda sem entrada de versao (template) — nada a comparar`);
-      }
-    } else {
-      // Nao assumir que o topo e a versao mais alta: comparar contra o maximo e
-      // avisar se a ordenacao estiver invertida (um CHANGELOG ascendente dava falso positivo).
-      // Precedencia SemVer (spec §11). Uma versao anterior desta funcao usava
-      // `split(/[-+]/, 2)` e comparacao de strings, o que dava tres bugs de uma vez:
-      // `beta.10 < beta.9` (string), `1.2.3+build < 1.2.3` (build tratado como
-      // pre-release) e `beta-9 == beta-2` (o split truncava o resto). Isso fazia o guard
-      // BLOQUEAR um CHANGELOG correctamente ordenado.
-      const parseV = (v) => {
-        const m = /^(\d+)\.(\d+)\.(\d+)(?:-([0-9A-Za-z.-]+))?(?:\+[0-9A-Za-z.-]+)?$/.exec(v);
-        if (!m) return null;
-        return { core: [+m[1], +m[2], +m[3]], pre: m[4] ? m[4].split(".") : [] };
-      };
-      const cmp = (a, b) => {
-        const pa = parseV(a);
-        const pb = parseV(b);
-        if (!pa || !pb) return a < b ? -1 : a > b ? 1 : 0;
-        for (let i = 0; i < 3; i++) if (pa.core[i] !== pb.core[i]) return pa.core[i] - pb.core[i];
-        // Build metadata ja foi descartado: nao conta para precedencia (spec §10).
-        if (pa.pre.length === 0 && pb.pre.length === 0) return 0;
-        if (pa.pre.length === 0) return 1;  // release > pre-release
-        if (pb.pre.length === 0) return -1;
-        for (let i = 0; i < Math.max(pa.pre.length, pb.pre.length); i++) {
-          const x = pa.pre[i];
-          const y = pb.pre[i];
-          if (x === undefined) return -1;   // menos identificadores = menor
-          if (y === undefined) return 1;
-          const nx = /^\d+$/.test(x);
-          const ny = /^\d+$/.test(y);
-          if (nx && ny) { if (+x !== +y) return +x - +y; continue; }
-          if (nx !== ny) return nx ? -1 : 1; // numerico < alfanumerico
-          if (x !== y) return x < y ? -1 : 1;
-        }
-        return 0;
-      };
-      const max = [...found].sort(cmp).at(-1);
-      if (cmp(found[0], max) !== 0) {
-        warn(`${CHANGELOG_PATH}: a entrada no topo e v${found[0]} mas a maior e v${max} — ordenar por versao decrescente`);
-      }
-      // Comparar por precedencia, nao por string: `1.2.3` e `1.2.3+build.5` sao a mesma
-      // versao (spec §10), e a igualdade de string reportava-as como divergentes.
-      const clean = pkgVersion.replace(/^v/, "");
-      if (cmp(clean, max) === 0) ok(`CHANGELOG (v${max}) === package.json`);
-      else warn(`package.json (${pkgVersion}) != maior versao do CHANGELOG (v${max}) — atualizar o CHANGELOG antes do commit`);
-    }
-  }
-  guardsRun++;
-}
-
+// Extraido para `guards/versions.mjs`.
+guardsRun += guardChangelogVersion({ read, warn, ok, note, skip });
 // --- Guard 4: termos obsoletos/banidos nos docs vivos ---
 // Adicionar entradas conforme forem renomeados ficheiros/termos ou mudarem contagens.
 // Exclui historico (*-archive.md, CHANGELOG.md).
@@ -443,462 +365,16 @@ if (workflows && (claudeCmds || geminiCmds)) {
 }
 
 // --- Guard 11: sanidade do .claude/settings.json ---
-// E a fronteira de seguranca do projeto e nao tinha rede nenhuma: um `allow` demasiado
-// largo passava CI sem sinal. Nao substitui revisao humana — apanha as regressoes obvias.
-const SETTINGS_PATH = ".claude/settings.json";
-const settingsRaw = read(SETTINGS_PATH);
-if (settingsRaw === null) {
-  skip(`Guard 11 (${SETTINGS_PATH}) — ficheiro ausente (projeto pode nao usar Claude Code)`);
-} else {
-  let settings = null;
-  try {
-    settings = JSON.parse(settingsRaw);
-  } catch (err) {
-    warn(`${SETTINGS_PATH} nao e JSON valido (${err instanceof Error ? err.message : String(err)})`);
-  }
-  if (settings) {
-    const perms = settings.permissions ?? {};
-    const asList = (v, name) => {
-      if (v === undefined) return [];
-      if (Array.isArray(v)) return v;
-      warn(`${SETTINGS_PATH}: \`permissions.${name}\` devia ser um array`);
-      return [];
-    };
-    const denyRaw = asList(perms.deny, "deny");
-    const allow = asList(perms.allow, "allow");
-    let issues = 0;
-    const flag = (msg) => {
-      warn(`${SETTINGS_PATH}: ${msg}`);
-      issues++;
-    };
-    // Filtrar DEPOIS de `flag` existir: uma versao anterior usava `warn` aqui, logo
-    // `issues` ficava 0 e o guard imprimia o WARN e o OK ao mesmo tempo.
-    const deny = denyRaw.filter((r) => {
-      if (typeof r === "string") return true;
-      flag(`entrada do \`deny\` que nao e string (${JSON.stringify(r)})`);
-      return false;
-    });
+// Extraido para `guards/settings.mjs` (eram 266 linhas). Os avisos vivem la, logo esse
+// ficheiro esta registado em `PARES` no mutation-sweep.mjs.
+guardsRun += guardSettings({ read, warn, ok, note, skip });
 
-    // Cobertura do deny de secrets: verificar que ALGUM padrao cobre cada caminho tipico,
-    // em vez de exigir os 4 literais. Um projeto com `Read(./**/.env*)` (superset estrito)
-    // era reportado como tendo o deny em falta.
-    const denyCovers = (target) =>
-      deny.some((rule) => {
-        const m = /^Read\((.*)\)$/.exec(rule);
-        if (!m) return false;
-        // Traducao glob->regex numa passagem unica. Fazer os `replace` em cadeia era
-        // um bug: o `.*` inserido pelo passo do `**` continha um `*` que o passo
-        // seguinte voltava a reescrever, produzindo `(?:.[^/]*/)?` — que nao casa com
-        // caminhos aninhados, e o guard reportava um deny correcto como estando em falta.
-        const glob = m[1].replace(/^\.\//, "");
-        let rxSrc = "";
-        for (let i = 0; i < glob.length; i++) {
-          if (glob[i] === "*" && glob[i + 1] === "*") {
-            if (glob[i + 2] === "/") { rxSrc += "(?:[^/]+/)*"; i += 2; }
-            else { rxSrc += ".*"; i += 1; }
-          } else if (glob[i] === "*") {
-            rxSrc += "[^/]*";
-          } else {
-            rxSrc += glob[i].replace(/[.+^${}()|[\]\\?]/, "\\$&");
-          }
-        }
-        const rx = new RegExp(`^${rxSrc}$`);
-        return rx.test(target);
-      });
-    // Alvos de TODAS as classes de secret que o template protege — nao so `.env*`.
-    // Apagar as regras de `*.pem`/`id_rsa`/`secrets/**` era invisivel.
-    for (const target of [
-      ".env", ".env.local", ".env.production", "a/b/.env", "a/.env.local",
-      "chave.pem", "a/b/tls.key", "id_rsa", "a/id_rsa.pub", ".npmrc",
-      "credentials.json", "a/secrets/db.json",
-    ]) {
-      if (!denyCovers(target)) flag(`nenhuma regra \`deny\` cobre a leitura de \`${target}\` (nota: o tradutor de glob nao expande braces \`{a,b}\`)`);
-    }
-
-    // Um `allow` demasiado largo anula o deny ao lado. Casos que ja foram reais aqui:
-    // `Bash(npm run lint*)` cobria `npm run lint-and-deploy`; `Bash(node .agent/scripts/*)`
-    // cobria qualquer ficheiro nesse caminho; `Bash(*)` e `Bash(rm -rf *)` abriam tudo; e
-    // `Bash` sem parenteses e a concessao maxima possivel.
-    // A propriedade que interessa nao e a FORMA da regra — e se o argumento fica
-    // constrangido. Uma versao anterior testava a forma e por isso aprovava
-    // `Bash(/bin/sh:*)` (prefixo derrota o `^`), `Bash(npm:*)` (npm exec e arbitrario),
-    // `Bash ` (sem trim), `Grep`/`NotebookEdit`/`mcp__*` (fora da lista hardcoded), e
-    // ao mesmo tempo marcava como perigosa a concessao estreita e legitima
-    // `Bash(node .agent/scripts/x.mjs:*)`.
-    // O que interessa e se o argumento fica constrangido. Formas que versoes anteriores
-    // aprovaram, cada uma por uma razao diferente:
-    //   `Bash(sh:*)`            -> apanhado
-    //   `Bash(/bin/sh:*)`       -> prefixo derrotava um `^` ancorado
-    //   `Bash(env -i sh:*)`     -> remover o wrapper deixava a FLAG como head
-    //   `Bash(BASH -c:*)`       -> comparacao case-sensitive
-    //   `Bash(git log; sh:*)`   -> metacaracteres de shell nunca eram olhados
-    //   `Read(~/**)`            -> so padroes feitos de `*` e `/` eram vistos
-    //   `WebFetch(*)`           -> so os nomes NUS eram cobertos
-    const WRAPPERS = new Set(["env", "command", "nohup", "nice", "time", "sudo", "doas", "xargs", "exec", "setsid", "stdbuf"]);
-    // Familia de interpretadores por PADRAO, nao por nome exacto: uma lista fechada
-    // deixava passar `pwsh`, `python3.12`, `ruby2.7`, `tclsh`, `bash5`.
-    const INTERP = /^(sh|bash|zsh|fish|dash|csh|ksh|tcsh|pwsh|powershell|node|deno|bun|python|ruby|perl|php|lua|tclsh|osascript|expect|rscript|gcc|clang)[\d.]*$/i;
-    // Subcomandos que reabrem execucao arbitraria apesar de o 2o token ser "fixo".
-    const REABRE = new Set([
-      "npm exec", "npm x", "npx", "pnpm exec", "pnpm dlx", "yarn dlx", "bun x", "bunx",
-      "deno run", "deno eval", "pip install", "uv run", "uvx", "git commit", "git -c",
-      "gh api", "make -f", "docker run", "kubectl exec",
-    ]);
-    const RISKY = new Set([
-      "sh", "bash", "zsh", "fish", "dash", "csh", "ksh", "tcsh",
-      "node", "deno", "bun", "python", "python3", "ruby", "perl", "php", "osascript", "lua",
-      "eval", "npm", "npx", "pnpm", "yarn", "make", "git", "find", "awk", "sed", "xargs",
-      "rm", "mv", "dd", "chmod", "chown", "curl", "wget", "nc", "ncat", "ssh", "scp",
-      "docker", "kubectl", "security", "defaults", "launchctl", "systemctl", "sudo", "doas",
-    ]);
-    // `;` `&&` `||` `|` backtick `$(` `>` `<` `\n` — encadeiam um segundo comando.
-    const SHELL_META = /[;|&`\n]|\$\(|>|</;
-    // Comandos que, mesmo com argumentos fixos, nao devem ser pre-aprovados.
-    const DESTRUCTIVE = new Set(["rm", "dd", "mv", "chmod", "chown", "mkfs", "shutdown", "reboot", "killall"]);
-    const baseName = (tok) => tok.split("/").pop().toLowerCase();
-
-    for (const raw of allow) {
-      if (typeof raw !== "string") {
-        flag(`entrada do \`allow\` que nao e string (${JSON.stringify(raw)}) — ignorada em silencio pelo motor`);
-        continue;
-      }
-      const rule = raw.trim();
-
-      // Nome de ferramenta sem `(...)`: concede a ferramenta INTEIRA.
-      if (/^[A-Za-z_][A-Za-z0-9_]*$/.test(rule)) {
-        flag(`\`${rule}\` sem \`(...)\` concede a ferramenta INTEIRA — enumerar os comandos/caminhos exatos`);
-        continue;
-      }
-
-      const parsed = /^([A-Za-z_][A-Za-z0-9_]*)\((.*)\)$/.exec(rule);
-      if (!parsed) {
-        flag(`\`${rule}\` nao tem a forma \`Ferramenta(padrao)\` — o motor pode ignora-la`);
-        continue;
-      }
-      const [, tool, bodyRaw] = parsed;
-      const body = bodyRaw.trim();
-
-      if (tool === "Bash") {
-        if (SHELL_META.test(body)) {
-          flag(`\`${rule}\` contem metacaracteres de shell — encadeia um segundo comando`);
-          continue;
-        }
-        const suffix = /^(.*[^*]):\*$/.exec(body);
-        if (body.includes("*") && !suffix) {
-          flag(`\`${rule}\` tem wildcard sem delimitador — enumerar os comandos exatos`);
-          continue;
-        }
-        // Procurar um comando perigoso em QUALQUER posicao do prefixo fixo, nao so na
-        // primeira. Descartar apenas o head nao bastava: em `sudo -u root sh:*` o loop
-        // consumia `sudo` e `-u`, e o head passava a ser `root` — o argumento do wrapper.
-        const fixed = suffix ? suffix[1] : body;
-        const tokens = fixed.split(/\s+/).filter(Boolean);
-        // Caminho fora do projeto em QUALQUER token, seja o comando fixo ou com `:*`.
-        const foraDoProjeto = tokens.find((t) => t.startsWith("~") || t.startsWith("/etc") || /(^|\/)\.\.(\/|$)/.test(t) || /^\/(Users|home|root|var|private)\//.test(t));
-        if (foraDoProjeto) {
-          flag(`\`${rule}\` referencia \`${foraDoProjeto}\`, fora do projeto`);
-          continue;
-        }
-        // Subcomando que reabre execucao: `npm:*` era marcado porque "npm exec e
-        // arbitrario", mas `npm exec:*` — que concede exactamente isso — passava.
-        const doisPrimeiros = tokens.slice(0, 2).map(baseName).join(" ");
-        if (suffix && (REABRE.has(doisPrimeiros) || REABRE.has(baseName(tokens[0] ?? "")))) {
-          flag(`\`${rule}\` pre-aprova \`${doisPrimeiros}\`, que reabre execucao arbitraria`);
-          continue;
-        }
-        // Um INTERPRETADOR em qualquer posicao e o caso mais grave: so e aceitavel com um
-        // alvo fixo que nao seja ele proprio um interpretador. Tratar wrappers como
-        // "comando perigoso com alvo fixo" era um buraco: em `env sh:*` o `env` ficava
-        // como head e o `sh` passava por alvo — concedendo shell arbitraria.
-        const interpAt = tokens.findIndex((t) => INTERP.test(baseName(t)));
-        if (interpAt !== -1) {
-          const alvo = tokens[interpAt + 1];
-          const alvoOk = alvo && !alvo.startsWith("-") && !alvo.includes("*") && !INTERP.test(baseName(alvo));
-          if (!alvoOk) {
-            flag(`\`${rule}\` deixa \`${baseName(tokens[interpAt])}\` receber argumentos livres — equivale a execucao arbitraria`);
-            continue;
-          }
-        }
-        const isRisky = (t) => RISKY.has(baseName(t)) || WRAPPERS.has(baseName(t));
-        const riskyAt = tokens.findIndex(isRisky);
-        const head = riskyAt === -1 ? "" : baseName(tokens[riskyAt]);
-        const next = riskyAt === -1 ? undefined : tokens[riskyAt + 1];
-        if (!suffix) {
-          // Comando FIXO: nao ha argumentos livres, logo `npx tsc --noEmit` e seguro.
-          // So interessam dois casos: um interpretador sozinho (shell interactiva) ou
-          // um comando destrutivo pre-aprovado.
-          if (riskyAt !== -1 && (next === undefined || DESTRUCTIVE.has(head))) {
-            flag(`\`${rule}\` pre-aprova \`${head}\` — shell interactiva ou comando destrutivo`);
-          }
-          continue;
-        }
-        if (riskyAt !== -1) {
-          // Com `:*` ha argumentos livres: so aceitavel se o comando perigoso vier
-          // seguido de um alvo FIXO (`node caminho/script.mjs:*`, `npm run lint:*`).
-          const alvoFixo = next && !next.startsWith("-") && !next.includes("*");
-          if (!alvoFixo) {
-            flag(`\`${rule}\` deixa \`${head}\` receber argumentos livres — equivale a execucao arbitraria`);
-            continue;
-          }
-        }
-        if (tokens.length === 1) {
-          // Um so token e desconhecido: largo mas nao necessariamente perigoso
-          // (`ls:*`, `cat:*`). NOTE, nao WARN — nao trava o gate. Interpretadores e
-          // wrappers ja foram marcados acima; despromove-los a NOTE foi uma regressao.
-          note(`${SETTINGS_PATH}: \`${rule}\` pre-aprova quaisquer argumentos de \`${baseName(tokens[0])}\` — considerar fixar mais`);
-        }
-        continue;
-      }
-
-      if (["Read", "Edit", "Write", "Glob", "Grep", "NotebookEdit", "MultiEdit", "NotebookRead"].includes(tool)) {
-        const pat = body.replace(/^\.\//, "");
-        if (pat.startsWith("~") || pat.startsWith("/") || pat.split("/").includes("..")) {
-          flag(`\`${rule}\` sai do projeto (\`~\`, caminho absoluto ou \`..\`) — restringir ao repo`);
-        } else if (/^[*/]+$/.test(pat)) {
-          flag(`\`${rule}\` abrange o repo inteiro — restringir a um subcaminho`);
-        }
-        continue;
-      }
-
-      // Qualquer outra ferramenta (WebFetch, Task, mcp__*): um padrao de puro wildcard
-      // e concessao total.
-      if (/^[*/]*\*[*/]*$/.test(body)) {
-        flag(`\`${rule}\` concede \`${tool}\` sem restricao — enumerar os alvos permitidos`);
-      }
-    }
-
-    // Cobertura do `ask`: `process-rules.md` declara commit/push/merge e instalacao de
-    // dependencias como "perguntar primeiro". Apagar a lista inteira era invisivel.
-    const ask = asList(perms.ask, "ask");
-    const askCovers = (cmd) => ask.some((r) => {
-      const m = /^Bash\((.*?):?\*?\)$/.exec(typeof r === "string" ? r : "");
-      return m ? cmd.startsWith(m[1].trim()) : false;
-    });
-    for (const cmd of ["git commit", "git push", "npm install"]) {
-      if (!askCovers(cmd)) flag(`\`${cmd}\` nao esta em \`ask\` nem em \`deny\` — o CLAUDE.md declara-o como "perguntar primeiro"`);
-    }
-
-    // Um `allow` que caia dentro de um prefixo negado ou perguntado e contradicao
-    // silenciosa: o repo tem `git commit` em `ask`, e `Bash(git commit -m:*)` no allow
-    // passava sem sinal.
-    const prefixOf = (r) => /^Bash\((.*?):?\*?\)$/.exec(r)?.[1]?.trim() ?? null;
-    for (const a of allow) {
-      if (typeof a !== "string") continue;
-      const pa = prefixOf(a);
-      if (!pa) continue;
-      for (const [lista, nome] of [[deny, "deny"], [asList(perms.ask, "ask"), "ask"]]) {
-        for (const other of lista) {
-          if (typeof other !== "string") continue;
-          const po = prefixOf(other);
-          if (po && pa.startsWith(po)) {
-            flag(`\`${a}\` no allow cai dentro de \`${other}\` no ${nome} — contradicao`);
-          }
-        }
-      }
-    }
-
-    // `bypassPermissions` anula tudo o que esta acima.
-    const mode = perms.defaultMode;
-    if (mode && !["default", "acceptEdits", "plan"].includes(mode)) {
-      flag(`\`defaultMode: "${mode}"\` desliga a fronteira de permissoes deste ficheiro`);
-    }
-
-    if (issues === 0) ok(`${SETTINGS_PATH} (deny de secrets cobre .env*, allow sem concessoes largas)`);
-  }
-  guardsRun++;
-}
-
-// --- Guard 12: o tamanho da checklist de sync-docs e dado DERIVADO ---
-// A checklist vive num ficheiro; o seu tamanho ("N pontos") esta citado em prosa noutros
-// quatro. Nada os ligava: a checklist cresceu para 25 e as quatro citacoes ficaram em 24,
-// sem nenhum guard a notar. Aqui recalcula-se a partir da fonte e compara-se com cada
-// citacao — mesma logica dos contadores do backlog.
-const syncDocs = readMeaningful(".agent/rules/sync-docs.md");
-if (syncDocs) {
-  // 12a: a numeracao e o invariante que se quebra a mao ao inserir um ponto no meio.
-  const nums = [...syncDocs.matchAll(/^(\d+)\. \[ \]/gm)].map((m) => Number(m[1]));
-  if (nums.length === 0) {
-    warn(".agent/rules/sync-docs.md nao tem pontos de checklist (`N. [ ]`) — a checklist esta vazia ou mudou de formato");
-  } else {
-    const esperado = nums.map((_, i) => i + 1);
-    const desalinhado = nums.findIndex((n, i) => n !== esperado[i]);
-    if (desalinhado !== -1) {
-      warn(
-        `.agent/rules/sync-docs.md: numeracao da checklist quebrada — ` +
-          `o ponto #${desalinhado + 1} esta escrito como "${nums[desalinhado]}." ` +
-          `(sequencia lida: ${nums.join(", ")})`
-      );
-    } else {
-      ok(`checklist sync-docs numerada 1-${nums.length} sem saltos`);
-    }
-
-    // 12b: cada citacao em prosa tem de igualar a contagem real.
-    const total = nums.length;
-    const alvos = [
-      ...(listDir(".agent/rules", ".md") || []).map((f) => `.agent/rules/${f}.md`),
-      ...(listDir(".agent/workflows", ".md") || []).map((f) => `.agent/workflows/${f}.md`),
-      "src/docs/agent-guide.md",
-      "CLAUDE.md",
-      "GEMINI.md",
-      "AGENTS.md",
-      "README.md",
-      "CONTRIBUTING.md",
-    ];
-    let citacoes = 0;
-    let erradas = 0;
-    for (const f of alvos) {
-      const c = read(f);
-      if (c === null) continue;
-      // "(N pontos" tambem descreve outras contagens (ex: o relatorio de fecho de sprint
-      // tem 6). So conta como citacao DESTA checklist se a mesma linha falar de sync-docs.
-      for (const linha of c.split("\n")) {
-        if (!linha.includes("sync-docs")) continue;
-        const m = /\((\d+)\s+pontos/.exec(linha);
-        if (!m) continue;
-        citacoes++;
-        if (Number(m[1]) !== total) {
-          warn(`${f} diz "(${m[1]} pontos" mas a checklist de sync-docs.md tem ${total}`);
-          erradas++;
-        }
-      }
-    }
-    if (citacoes === 0) {
-      warn(`nenhum ficheiro cita o tamanho da checklist ("(N pontos") — o ponteiro obrigatorio em process-rules.md desapareceu?`);
-    } else if (erradas === 0) {
-      ok(`${citacoes} citacao(oes) de "(N pontos" coerentes com ${total}`);
-    }
-  }
-  guardsRun++;
-} else {
-  skip(`Guard 12 — ${why(".agent/rules/sync-docs.md")}`);
-}
-
-// --- Guard 12c: o numero de fases do metodo por ticket e dado DERIVADO ---
-// Mesmo defeito do 12b, encontrado a olho: o ficheiro tem 6 fases (0 a 5) e tres sitios
-// diziam "5 fases", enumerando cinco. A que faltava em todos era a Fase 4 (o leitor
-// independente) — a mais cara, e portanto a que mais convem esquecer. Num ficheiro que diz
-// "nenhuma se salta", omitir uma dos resumos e pior do que uma gralha.
-const metodo = readMeaningful(".agent/rules/ticket-method.md");
-if (metodo) {
-  const fases = [...metodo.matchAll(/^##\s*Fase\s+(\d+)\b/gm)].map((m) => Number(m[1]));
-  if (fases.length === 0) {
-    warn(".agent/rules/ticket-method.md nao tem cabecalhos `## Fase N` — o metodo mudou de formato?");
-  } else {
-    // A numeracao comeca em 0 e nao deve ter saltos: uma fase renumerada a mao quebra as
-    // referencias "Fase N" espalhadas pelos workflows.
-    const esperado = fases.map((_, i) => i);
-    const mau = fases.findIndex((n, i) => n !== esperado[i]);
-    if (mau !== -1) {
-      warn(
-        `.agent/rules/ticket-method.md: fases numeradas [${fases.join(", ")}] — esperado 0..${fases.length - 1} sem saltos`
-      );
-    }
-
-    const total = fases.length;
-    const alvos = [
-      ...(listDir(".agent/rules", ".md") || []).map((f) => `.agent/rules/${f}.md`),
-      ...(listDir(".agent/workflows", ".md") || []).map((f) => `.agent/workflows/${f}.md`),
-      "src/docs/agent-guide.md",
-      "README.md",
-      "AGENTS.md",
-    ];
-    // Formas em que o total aparece: "6 fases", "as 6 fases", "Seis fases".
-    const PALAVRA = { 3: "tres", 4: "quatro", 5: "cinco", 6: "seis", 7: "sete", 8: "oito" };
-    let citacoes = 0;
-    let erradas = 0;
-    for (const f of alvos) {
-      const c = read(f);
-      if (c === null) continue;
-      for (const linha of c.split("\n")) {
-        const digito = /(\d+)\s+fases/i.exec(linha);
-        const palavra = /\b(tres|quatro|cinco|seis|sete|oito)\s+fases/i.exec(linha);
-        if (!digito && !palavra) continue;
-        citacoes++;
-        const escrito = digito ? Number(digito[1]) : null;
-        const esperadoPalavra = PALAVRA[total];
-        if (digito && escrito !== total) {
-          warn(`${f} diz "${escrito} fases" mas ${".agent/rules/ticket-method.md"} tem ${total}`);
-          erradas++;
-        } else if (palavra && palavra[1].toLowerCase() !== esperadoPalavra) {
-          warn(`${f} diz "${palavra[1]} fases" mas o metodo tem ${total} (${esperadoPalavra})`);
-          erradas++;
-        }
-      }
-    }
-    if (citacoes === 0) {
-      warn(`nenhum ficheiro cita o numero de fases do metodo — o ponteiro obrigatorio em process-rules.md desapareceu?`);
-    } else if (erradas === 0) {
-      ok(`${citacoes} citacao(oes) de "N fases" coerentes com ${total}`);
-    }
-  }
-  guardsRun++;
-} else {
-  skip(`Guard 12c — ${why(".agent/rules/ticket-method.md")}`);
-}
-
+// --- Guards 12 e 12c: contagens citadas em prosa como dados DERIVADOS ---
+// Extraidos para `guards/derived-counts.mjs`.
+guardsRun += guardDerivedCounts({ read, readMeaningful, warn, ok, skip, why, listDir });
 // --- Guards CONFIGURAVEIS: versoes de dependencias documentadas ---
-// pattern: regex que captura a versao no markdown (ex: "Next.js 16.2.2")
-const CHECKS = [
-  // { name: "Next.js", pkg: "next", pattern: /Next\.js\s+(\d+(?:\.\d+(?:\.\d+)?)?)/g, files: [".agent/rules/core-rules.md"] },
-];
-
-function cleanVersion(raw) {
-  return raw.replace(/^[\^~>=<\s]*/g, "");
-}
-function parts(v) {
-  const m = cleanVersion(v).match(/^(\d+)(?:\.(\d+))?(?:\.(\d+))?/);
-  if (!m) return null;
-  return { major: +m[1], minor: m[2] != null ? +m[2] : null, patch: m[3] != null ? +m[3] : null };
-}
-function isOutdated(documented, actual) {
-  const d = parts(documented), a = parts(actual);
-  if (!d || !a) return false;
-  if (d.minor === null) return d.major !== a.major;
-  if (d.patch === null) return d.major !== a.major || d.minor !== a.minor;
-  return d.major !== a.major || d.minor !== a.minor || d.patch !== a.patch;
-}
-
-let pkgParsed = null;
-if (pkgRaw) {
-  try {
-    pkgParsed = JSON.parse(pkgRaw);
-  } catch {
-    /* Guard 3 ja avisou sobre package.json invalido */
-  }
-}
-if (CHECKS.length === 0) {
-  skip("Guards de versoes de dependencias — lista CHECKS vazia (opt-in)");
-} else if (!pkgParsed) {
-  skip("Guards de versoes de dependencias — sem package.json legivel");
-} else {
-  const pkg = pkgParsed;
-  for (const check of CHECKS) {
-    const dep = pkg.dependencies?.[check.pkg] || pkg.devDependencies?.[check.pkg];
-    if (!dep) {
-      skip(`${check.name} — nao esta no package.json`);
-      continue;
-    }
-    const actual = cleanVersion(dep);
-    for (const file of check.files) {
-      const content = read(file);
-      if (content === null) {
-        skip(`${file} — nao encontrado`);
-        continue;
-      }
-      const rx = new RegExp(check.pattern.source, check.pattern.flags.includes("g") ? check.pattern.flags : check.pattern.flags + "g");
-      const matches = [...content.matchAll(rx)];
-      if (matches.length === 0) {
-        ok(`${check.name} — sem referencia de versao em ${file}`);
-        continue;
-      }
-      for (const match of matches) {
-        if (isOutdated(match[1], actual)) warn(`${check.name} em ${file}: documentado ${match[1]}, atual ${actual}`);
-        else ok(`${check.name} = ${match[1]} (bate com ${actual})`);
-      }
-    }
-    guardsRun++;
-  }
-}
+// Extraidos para `guards/versions.mjs`. Configurar o `CHECKS` la.
+guardsRun += guardDependencyVersions({ read, warn, ok, skip });
 
 console.log("");
 console.log(`  ${guardsRun} guard(s) executado(s), ${guardsSkipped} saltado(s).`);
