+++ .agent/scripts/test-guards.mjs
@@ -1,282 +1,24 @@
+#!/usr/bin/env node
 /**
  * Testes dos Doc Guards — {{PROJECT_NAME}}
  *
- * Testes NEGATIVOS: para cada guard, quebra-se deliberadamente o que ele promete
- * verificar e afirma-se que avisa E sai != 0. Um guard que passa quando devia falhar
- * e pior que guard nenhum, porque produz confianca infundada — foi exatamente assim
- * que dois guards deste ficheiro ficaram anos sem apanhar nada.
+ * Entry point das suites dos doc guards. O harness (sandbox, sandbox sintetico, `test()`)
+ * vive em `test-harness.mjs`; os blocos maiores foram para modulos que espelham a divisao
+ * do verificador (`tests-settings.mjs`, `tests-derived-counts.mjs`).
  *
- * Sem dependencias e sem package.json (o template nao traz nenhum, por desenho):
- * corre com `node`, como os restantes scripts de `.agent/scripts/`.
+ * Correr APOS qualquer alteracao a `check-doc-versions.mjs` ou a `guards/*.mjs`: um guard
+ * que passa quando devia falhar produz confianca infundada. E depois a varredura de
+ * mutacao (`mutation-sweep.mjs`), porque esta suite ficar verde nao prova que afirma algo.
  *
- * Cada teste corre numa COPIA em tmp, nunca no repo. O guard ancora-se a raiz
- * derivada da sua propria localizacao, logo copiar o script para a sandbox faz com
- * que ele valide a sandbox — e por isso que os testes conseguem mutar os inputs.
- *
- * Uso:
  *   node .agent/scripts/test-guards.mjs
- *
- * Sai != 0 se algum teste falhar. Corre em cada push/PR no job `guard-tests` do ci.yml
- * (sem `package.json` e sem gate do `detect` — o template puro e exatamente o caso coberto).
  */
 
-import { cpSync, mkdtempSync, mkdirSync, rmSync, readFileSync, readdirSync, writeFileSync, appendFileSync, existsSync } from "fs";
-import { execFileSync } from "child_process";
-import { fileURLToPath } from "url";
-import { dirname, resolve, join } from "path";
-import { tmpdir } from "os";
-
-const ROOT = resolve(dirname(fileURLToPath(import.meta.url)), "..", "..");
-const GUARD = ".agent/scripts/check-doc-versions.mjs";
-
-// Conjunto minimo que os guards leem. Copiado a cada teste para isolar mutacoes.
-const FIXTURE_PATHS = [
-  ".agent/rules",
-  ".agent/workflows",
-  ".agent/context",
-  ".agent/scripts",
-  ".claude/commands",
-  ".claude/settings.json",
-  ".gemini/commands",
-  "CLAUDE.md",
-  "GEMINI.md",
-  "AGENTS.md",
-  ".nvmrc",
-  "src/docs/CHANGELOG.md",
-  "src/docs/agent-guide.md",
-];
-
-let passed = 0;
-const failures = [];
-
-function sandbox() {
-  const dir = mkdtempSync(join(tmpdir(), "guard-test-"));
-  for (const p of FIXTURE_PATHS) {
-    const src = join(ROOT, p);
-    if (!existsSync(src)) continue;
-    const dest = join(dir, p);
-    mkdirSync(dirname(dest), { recursive: true });
-    cpSync(src, dest, { recursive: true });
-  }
-  return dir;
-}
-
-function runGuard(dir, cwd) {
-  try {
-    // `stdio` explicito: o ramo de erro concatenava stdout+stderr e o de sucesso so
-    // devolvia stdout, logo um aviso em stderr com exit 0 era invisivel.
-    const stdout = execFileSync(process.execPath, [join(dir, GUARD)], {
-      stdio: ["ignore", "pipe", "pipe"],
-      // `cwd` e relativo a sandbox. Uma versao anterior passava `cwd ?? dir` com um
-      // `expect.cwd` sempre `undefined`, pelo que os testes de cwd corriam na raiz e
-      // eram copias exatas do baseline — passavam com o defeito reintroduzido.
-      cwd: cwd ? join(dir, cwd) : dir,
-      encoding: "utf8",
-    });
-    return { code: 0, out: stdout };
-  } catch (err) {
-    return { code: err.status ?? -1, out: (err.stdout ?? "") + (err.stderr ?? "") };
-  }
-}
-
-/**
- * @param name    descricao do cenario
- * @param mutate  (dir) => void — a quebra a aplicar; omitir para o baseline
- * @param expect  {
- *   code,                 exit esperado (0 ou 1)
- *   includes?: string[],  afirmado contra as linhas WARN quando code=1, senao contra o output
- *   anyOut?: string[],    afirmado sempre contra o output inteiro (linhas OK/SKIP/NOTE)
- *   excludes?: string[],  nao pode aparecer em nenhum sitio do output
- *   cwd?: string,         subdiretorio da sandbox de onde correr o guard
- *   synthetic?: boolean,  partir da fixture limpa por construcao em vez da copia do repo
- * }
- * `mutate` pode devolver { includes?, excludes? } extra, para expectativas calculadas.
- */
-function test(name, mutate, expect) {
-  // `synthetic: true` parte da fixture limpa por construcao, em vez de uma copia do
-  // repo. Necessario para testes cuja mutacao produz um aviso que o repo real pode ja
-  // ter — nesse caso o aviso nao seria "novo" e o teste media o estado do repo.
-  const dir = expect.synthetic ? syntheticSandbox() : sandbox();
-  const base = expect.synthetic ? syntheticBaselineWarns : baselineWarns;
-  try {
-    // Um `throw` no setup abortava o processo a meio: os testes seguintes nunca corriam
-    // e nao havia linha de resumo. O ficheiro irmao ja tinha esta guarda; este nao.
-    let extra = {};
-    if (mutate) {
-      try {
-        extra = mutate(dir) ?? {};
-      } catch (err) {
-        failures.push({ name, problems: [`setup rebentou: ${err.message}`], out: "" });
-        console.log(`  FAIL  ${name}`);
-        console.log(`          setup rebentou: ${err.message}`);
-        return;
-      }
-    }
-    const { code, out } = runGuard(dir, expect.cwd);
-    const problems = [];
-    if (!out.includes("=== Doc Guards ===")) {
-      problems.push(`o guard nao produziu output (rebentou?): ${out.slice(0, 160)}`);
-    }
-    // Afirmacao por DIFERENCA face ao baseline, nao pelo exit code absoluto. Uma versao
-    // anterior saltava os testes `code: 0` quando o repo tinha avisos — desligava 17 dos
-    // 64 em silencio, com exit 0 e a dizer "todos passaram". Agora todos correm sempre.
-    // INVARIANTE, verificada em cada corrida: o exit code tem de refletir os avisos.
-    // Sem isto, trocar a ultima linha do guard por `process.exit(1)` dava 84/84 verde —
-    // as assercoes diferenciais falam de mensagens e nao do veredicto.
-    const temWarn = out.includes("  WARN  ");
-    if (temWarn && code === 0) problems.push("imprimiu WARN mas saiu 0");
-    if (!temWarn && code !== 0) problems.push(`nao imprimiu WARN mas saiu ${code}`);
-    const novos = [...warnsOf(out)].filter((w) => !base.has(w));
-    if (expect.code === 0) {
-      if (novos.length) problems.push(`nao devia acrescentar avisos; acrescentou ${novos.length}: ${novos[0]}`);
-    } else {
-      if (novos.length === 0) problems.push("devia acrescentar pelo menos um aviso novo — nao acrescentou nenhum");
-      if (code === 0) problems.push("devia sair != 0");
-    }
-    // CAUSA-RAIZ de quatro rondas de defeitos: `out.includes(...)` nao olha ao NIVEL da
-    // linha. Um teste `code: 1` era satisfeito por uma linha `NOTE` com o mesmo texto, ou
-    // pelo WARN de OUTRO guard que a mutacao tambem disparava — e sabotar o guard sob teste
-    // ficava invisivel. Um teste que espera aviso afirma-se contra as linhas WARN e mais
-    // nada; um que espera sucesso pode afirmar OK/SKIP/NOTE, logo usa o output inteiro.
-    const alvo = expect.code === 0 ? out : out.split("\n").filter((l) => l.trimStart().startsWith("WARN")).join("\n");
-    const ondeAlvo = expect.code === 0 ? "output" : "linhas WARN";
-    for (const s of [...(expect.includes ?? []), ...(extra.includes ?? [])]) {
-      if (!alvo.includes(s)) problems.push(`${ondeAlvo} devia conter "${s}"`);
-    }
-    // `anyOut`: para afirmar linhas que NAO sao WARN (OK/SKIP/NOTE) num teste que
-    // ainda assim espera exit != 0.
-    for (const s of expect.anyOut ?? []) {
-      if (!out.includes(s)) problems.push(`output devia conter "${s}"`);
-    }
-    for (const s of [...(expect.excludes ?? []), ...(extra.excludes ?? [])]) {
-      if (out.includes(s)) problems.push(`output NAO devia conter "${s}"`);
-    }
-    if (problems.length) {
-      failures.push({ name, problems, out });
-      console.log(`  FAIL  ${name}`);
-      for (const p of problems) console.log(`          ${p}`);
-    } else {
-      passed++;
-      console.log(`  PASS  ${name}`);
-    }
-  } finally {
-    rmSync(dir, { recursive: true, force: true });
-  }
-}
-
-// Helpers de mutacao
-const file = (dir, p) => join(dir, p);
-const readF = (dir, p) => readFileSync(file(dir, p), "utf8");
-const writeF = (dir, p, s) => writeFileSync(file(dir, p), s);
-const patchSettings = (dir, fn) => {
-  const cfg = JSON.parse(readF(dir, ".claude/settings.json"));
-  fn(cfg);
-  writeF(dir, ".claude/settings.json", JSON.stringify(cfg, null, 2));
-};
-
-/** Linhas de tabela com o caminho de cada workflow, para um CLAUDE.md sintetico nao
- *  disparar o Guard 7 (que exige todos os workflows listados). */
-const listWorkflowRows = (dir) =>
-  readdirSync(join(dir, ".agent/workflows"))
-    .filter((f) => f.endsWith(".md"))
-    .map((f) => `| ${f.slice(0, -3)} | \`.agent/workflows/${f}\` |`);
-const dropLinesContaining = (dir, p, needle) =>
-  writeF(dir, p, readF(dir, p).split("\n").filter((l) => !l.includes(needle)).join("\n"));
-
-/** Repo minimo e LIMPO POR CONSTRUCAO, montado do zero — nao e copia do repo real.
- *  Existe porque as assercoes diferenciais nao conseguem afirmar o exit code: num
- *  baseline que ja avisa, "nao acrescentou avisos" e satisfeito por um guard que
- *  falha sempre. Aqui sabemos que o veredicto correcto e exit 0, logo podemos exigi-lo. */
-function syntheticSandbox() {
-  const dir = mkdtempSync(join(tmpdir(), "guard-synth-"));
-  const w = (rel, body) => {
-    const full = join(dir, rel);
-    mkdirSync(dirname(full), { recursive: true });
-    writeFileSync(full, body);
-  };
-  mkdirSync(join(dir, ".agent", "scripts"), { recursive: true });
-  cpSync(join(ROOT, GUARD), join(dir, GUARD));
-
-  for (const r of ["core-rules", "process-rules", "anti-patterns"]) {
-    w(`.agent/rules/${r}.md`, `# ${r}\n\nConteudo minimo.\n`);
-  }
-  w(".agent/workflows/plan.md", "# /plan\n");
-  w(".claude/commands/plan.md", "---\ndescription: x\n---\n\nLer `.agent/workflows/plan.md`.\n");
-  w(".gemini/commands/plan.toml", 'description = "x"\nprompt = "Le .agent/workflows/plan.md"\n');
-  w(".agent/context/session.md", "# Session\n");
-  const entry = [
-    "# Entry",
-    "",
-    "@.agent/rules/core-rules.md",
-    "@.agent/rules/process-rules.md",
-    "@.agent/rules/anti-patterns.md",
-    // Importadas mas NAO criadas: e o estado do template antes do bootstrap, e da aos
-    // guards 1 e 8 o `SKIP` que dois testes afirmam. SKIP nao e WARN, logo a fixture
-    // continua limpa.
-    "@.agent/rules/business-logic.md",
-    "@.agent/rules/pages-architecture.md",
-    "@.agent/context/session.md",
-    "",
-    "| Workflow | Ficheiro |",
-    "|---|---|",
-    "| Planear | `.agent/workflows/plan.md` |",
-  ].join("\n") + "\n";
-  w("CLAUDE.md", entry);
-  w("GEMINI.md", entry.replace(/^@(.*)$/gm, "@[$1]"));
-  w("AGENTS.md", "# Agents\n\nWorkflows: `plan`\n");
-  w("src/docs/agent-guide.md", "# Guia\n\n| **`/plan`** | Planear |\n");
-  w(".nvmrc", "24\n");
-  w(".claude/settings.json", JSON.stringify({
-    permissions: {
-      // Deny e ask completos: a fixture tem de satisfazer a cobertura que o Guard 11
-      // exige, senao deixa de estar "limpa por construcao".
-      deny: ["Read(./.env)", "Read(./.env.*)", "Read(./**/.env)", "Read(./**/.env.*)", "Read(./**/*.pem)", "Read(./**/*.key)", "Read(./**/id_rsa*)", "Read(./**/.npmrc)", "Read(./**/credentials*)", "Read(./**/secrets/**)"],
-      ask: ["Bash(git commit:*)", "Bash(git push:*)", "Bash(npm install:*)"],
-      allow: ["Read(./.agent/**)", "Bash(npx tsc --noEmit)"],
-    },
-  }, null, 2));
-  return dir;
-}
-
-console.log("\n=== Testes dos Doc Guards ===\n");
-
-// A sandbox e uma COPIA do repo real. Em vez de assumir que o repo esta limpo (ou de
-// saltar testes quando nao esta), captura-se o conjunto de WARN do baseline UMA vez e
-// tudo se afirma por diferenca. Assim os 64 testes correm sempre e o veredicto e sobre
-// o guard, nao sobre o estado do repo.
-const warnsOf = (out) =>
-  new Set(out.split("\n").filter((l) => l.trimStart().startsWith("WARN")).map((l) => l.trim()));
-let baselineWarns;
-let syntheticBaselineWarns;
-{
-  const dir = sandbox();
-  try {
-    const base = runGuard(dir);
-    if (!base.out.includes("=== Doc Guards ===")) {
-      console.log("  ERRO  o guard nao produziu output no baseline — abortar:\n" + base.out);
-      process.exit(1);
-    }
-    baselineWarns = warnsOf(base.out);
-  } finally {
-    rmSync(dir, { recursive: true, force: true });
-  }
-  const synth = syntheticSandbox();
-  try {
-    const baseSynth = runGuard(synth);
-    if (!baseSynth.out.includes("=== Doc Guards ===")) {
-      console.log("  ERRO  o guard nao produziu output na fixture sintetica — abortar:\n" + baseSynth.out);
-      process.exit(1);
-    }
-    syntheticBaselineWarns = warnsOf(baseSynth.out);
-  } finally {
-    rmSync(synth, { recursive: true, force: true });
-  }
-  if (baselineWarns.size > 0) {
-    console.log(`  NOTA  o repo tem ${baselineWarns.size} aviso(s) no baseline dos doc guards.`);
-    console.log("        Todos os testes correm: as assercoes sao por DIFERENCA face a esse baseline.");
-    console.log("        Corre `node .agent/scripts/check-doc-versions.mjs` para os ver.\n");
-  }
-}
+import { rmSync, mkdirSync, writeFileSync, readFileSync, appendFileSync, existsSync } from "fs";
+import { join, dirname } from "path";
+import { test, sandbox, syntheticSandbox, runGuard, file, readF, writeF, patchSettings,
+         listWorkflowRows, dropLinesContaining, GUARD, GUARD_MODULES, ROOT, resumo, registarResultado } from "./test-harness.mjs";
+import { registar as registarSettings } from "./tests-settings.mjs";
+import { registar as registarDerivedCounts } from "./tests-derived-counts.mjs";
 
 // --- Baseline -----------------------------------------------------------------
 // Estes dois nao usam `test()`: correm contra a fixture sintetica, nao contra o repo.
@@ -291,14 +33,7 @@ let syntheticBaselineWarns;
     }
     if (out.includes("  WARN  ")) problems.push(`nao devia haver WARN: ${out.split("\n").find((l) => l.includes("  WARN  "))}`);
     const name = "sintetico: repo limpo por construcao sai 0 e declara que passou";
-    if (problems.length) {
-      failures.push({ name, problems, out });
-      console.log(`  FAIL  ${name}`);
-      for (const p of problems) console.log(`          ${p}`);
-    } else {
-      passed++;
-      console.log(`  PASS  ${name}`);
-    }
+    registarResultado(name, problems, out);
   } finally {
     rmSync(dir, { recursive: true, force: true });
   }
@@ -314,14 +49,7 @@ let syntheticBaselineWarns;
     const problems = [];
     if (code === 0) problems.push("devia sair != 0");
     if (!out.includes("  WARN  ")) problems.push("devia imprimir WARN");
-    if (problems.length) {
-      failures.push({ name, problems, out });
-      console.log(`  FAIL  ${name}`);
-      for (const p of problems) console.log(`          ${p}`);
-    } else {
-      passed++;
-      console.log(`  PASS  ${name}`);
-    }
+    registarResultado(name, problems, out);
   } finally {
     rmSync(dir, { recursive: true, force: true });
   }
@@ -449,6 +177,9 @@ test("G4: termo banido e apanhado em TODOS os ficheiros, nao so no primeiro", (d
     "const BANNED = [\n",
     'const BANNED = [\n  { re: /Fronteiras/g, msg: "termo de teste" },\n'
   );
+  // Um `.replace()` que nao casa devolve o ficheiro intacto, e o teste passaria pela razao
+  // errada. Aconteceu com o `CHECKS` quando ele mudou de ficheiro: melhor rebentar aqui.
+  if (g === readF(dir, GUARD)) throw new Error("nao encontrei 'const BANNED = [' no GUARD");
   writeF(dir, GUARD, g);
 }, { code: 1, includes: ["CLAUDE.md:", "GEMINI.md:", "termo de teste"] });
 
@@ -561,174 +292,8 @@ test("G10: wrapper vazio avisa", (dir) => {
   writeF(dir, ".gemini/commands/debug.toml", "");
 }, { code: 1, includes: [".gemini/commands/debug.toml nao aponta"] });
 
-// --- Guard 11: sanidade do settings.json -------------------------------------
-test("G11: deny de .env retirado por completo avisa", (dir) => {
-  // Escrever primeiro um settings.json com o deny COMPLETO: num projeto derivado que
-  // ja nao o tenha, o aviso estaria no baseline e a assercao diferencial nao o veria.
-  writeF(dir, ".claude/settings.json", JSON.stringify({
-    permissions: { deny: ["Read(./.env)", "Read(./**/.env)", "Read(./.env.*)", "Read(./**/.env.*)"], allow: [] },
-  }, null, 2));
-  const cfg = JSON.parse(readF(dir, ".claude/settings.json"));
-  cfg.permissions.deny = [];
-  writeF(dir, ".claude/settings.json", JSON.stringify(cfg, null, 2));
-}, { code: 1, includes: ["nenhuma regra `deny` cobre a leitura de"] });
 
-test("G11: wildcard sem delimitador no allow avisa", (dir) => {
-  const cfg = JSON.parse(readF(dir, ".claude/settings.json"));
-  cfg.permissions.allow.push("Bash(node .agent/scripts/*)", "Bash(npm run lint*)");
-  writeF(dir, ".claude/settings.json", JSON.stringify(cfg, null, 2));
-}, { code: 1, includes: ["Bash(node .agent/scripts/*)", "Bash(npm run lint*)", "wildcard sem delimitador"] });
-
-test("G11: `:*` e delimitador valido, nao avisa", (dir) => {
-  const cfg = JSON.parse(readF(dir, ".claude/settings.json"));
-  cfg.permissions.allow.push("Bash(git log:*)");
-  writeF(dir, ".claude/settings.json", JSON.stringify(cfg, null, 2));
-}, { code: 0, excludes: ["wildcard sem delimitador"] });
-
-test("G11: settings.json invalido avisa", (dir) => {
-  writeF(dir, ".claude/settings.json", "{ nope,, }");
-}, { code: 1, includes: ["nao e JSON valido"] });
-
-test("G11: settings.json ausente da SKIP visivel", (dir) => {
-  rmSync(file(dir, ".claude/settings.json"));
-}, { code: 0, includes: ["SKIP  Guard 11"] });
-
-// --- Guard 11: formas de bypass que uma versao anterior aprovava ---------------
-// Estas existem porque a primeira versao do Guard 11 so olhava para o `*` precedido de
-// caractere de palavra. `Bash(*)`, `Bash` sem parenteses e `Bash(sh:*)` passavam todos.
-test("G11: `Bash(*)` e apanhado", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash(*)"));
-}, { code: 1, includes: ["Bash(*)", "wildcard sem delimitador"] });
-
-test("G11: `Bash` sem parenteses (concessao maxima) e apanhado", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash"));
-}, { code: 1, includes: ["concede a ferramenta INTEIRA"] });
-
-test("G11: `Bash(sh:*)` respeita a forma `:*` mas e execucao arbitraria", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash(sh:*)"));
-}, { code: 1, includes: ["receber argumentos livres"] });
-
-test("G11: `Bash(node:*)` idem", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash(node:*)"));
-}, { code: 1, includes: ["receber argumentos livres"] });
-
-test("G11: `Read(./**)` abrange o repo inteiro", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Read(./**)"));
-}, { code: 1, includes: ["abrange o repo inteiro"] });
-
-test("G11: `defaultMode: bypassPermissions` desliga tudo", (dir) => {
-  patchSettings(dir, (c) => (c.permissions.defaultMode = "bypassPermissions"));
-}, { code: 1, includes: ["desliga a fronteira de permissoes"] });
-
-test("G11: `allow` que nao e array nao rebenta", (dir) => {
-  patchSettings(dir, (c) => (c.permissions.allow = {}));
-}, { code: 1, includes: ["devia ser um array"], excludes: ["is not iterable"] });
-
-test("G11: deny mais ESTRITO nao e falso positivo", (dir) => {
-  // Supersets estritos dos padroes do template: um por classe de secret.
-  patchSettings(dir, (c) => (c.permissions.deny = [
-    "Read(./.env*)", "Read(./**/.env*)", "Read(./**/*.pem)", "Read(./**/*.key)",
-    "Read(./**/id_rsa*)", "Read(./**/.npmrc)", "Read(./**/credentials*)", "Read(./**/secrets/**)",
-  ]));
-}, { code: 0, excludes: ["nenhuma regra `deny` cobre"] });
-
-// --- Guard 11: bypasses por PREFIXO e nomes de ferramenta fora da lista -------
-// Todos estes passavam a verde quando o guard testava a FORMA da regra em vez da
-// propriedade ("o argumento fica constrangido?").
-for (const rule of [
-  "Bash(/bin/sh:*)",        // prefixo derrota um `^` ancorado
-  "Bash(/bin/bash -c:*)",   // interpretador + flag = execucao arbitraria
-  "Bash(/usr/bin/env sh:*)",// wrapper `env` a esconder o interpretador
-  "Bash(npm:*)",            // npm exec -- qualquer coisa
-  "Bash(npx:*)",
-  "Bash(git:*)",            // git -c core.pager='sh -c ...'
-  "Bash(find:*)",           // find -exec
-  "Bash(awk:*)",            // awk 'BEGIN{system(...)}'
-  "Bash(docker:*)",
-  "Bash ",                  // sem trim, era um nome de ferramenta nu
-  "Grep",                   // fora da lista hardcoded de 6 nomes
-  "NotebookEdit",           // ferramenta de ESCRITA fora da lista
-  "mcp__servidor__tool",    // ferramentas MCP nao eram olhadas
-  "Read(./*)",              // raiz inteira
-  // Ronda 4: cada uma passava por uma razao diferente.
-  "Bash(env -i sh:*)",      // remover o wrapper deixava a FLAG como head
-  "Bash(sudo -u root sh:*)",// o head passava a ser `root`, argumento do wrapper
-  "Bash(nice -n 0 bash:*)",
-  "Bash(xargs -I{} sh:*)",
-  "Bash(BASH -c:*)",        // comparacao case-sensitive
-  "Bash(/bin/SH -c:*)",
-  "Bash(git log; sh:*)",    // metacaracteres de shell nunca eram olhados
-  "Bash(git log|sh:*)",
-  "Bash(cd / && sh:*)",
-  "Bash(sh)",               // comando fixo: shell interactiva
-  "Bash(rm -rf ~)",         // comando fixo destrutivo
-  "Read(~/**)",             // home inteira, inclui ~/.ssh
-  "Read(/Users/**)",        // caminho absoluto
-  "Read(../**)",            // sai do repo
-  "Write(~/.ssh/**)",
-  "WebFetch(*)",            // ferramenta com parenteses fora de Bash/ficheiro
-  "Task(*)",
-  "mcp__srv__tool(*)",
-]) {
-  test(`G11: \`${rule}\` e apanhado`, (dir) => {
-    patchSettings(dir, (c) => c.permissions.allow.push(rule));
-    // Afirmar que foi ESTA regra a ser marcada. Sem isto o teste passava com o guard a
-    // avisar por qualquer outra razao — vermelho, mas nao na assercao certa.
-    // `.trim()`: o guard reporta a regra normalizada (o caso `"Bash "` perde o espaco).
-    return { includes: [rule.trim()] };
-  }, { code: 1 });
-}
-
-test("G11: entrada do allow que nao e string e apanhada", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push({ tool: "Bash", args: "*" }));
-}, { code: 1, includes: ["nao e string"] });
-
-// E o simetrico: concessoes ESTREITAS e legitimas nao podem ser falsos positivos.
-for (const rule of [
-  "Bash(node .agent/scripts/check-doc-versions.mjs:*)", // interpretador com alvo FIXO
-  "Bash(npm run lint:*)",                               // prefixo de 3 tokens
-  "Bash(git log:*)",
-  "Read(./.agent/**)",
-  // Ronda 4: comandos fixos nao tem argumentos livres, logo sao seguros.
-  "Bash(npx tsc --noEmit)",
-  "Bash(git status)",
-  "Glob(**/*.ts)",
-  // Read-only com um so token: NOTE, nao WARN — nao pode travar o gate.
-  "Bash(ls:*)",
-  "Bash(cat:*)",
-]) {
-  test(`G11: \`${rule}\` NAO e falso positivo`, (dir) => {
-    patchSettings(dir, (c) => c.permissions.allow.push(rule));
-  }, { code: 0 });
-}
-
-// Ramos acrescentados na ronda 5 — descobertos pela varredura de mutacao, que revelou
-// que eu os tinha escrito sem teste (o proprio AP1 a acontecer).
-test("G11: caminho fora do projeto e apanhado", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash(cat ~/.ssh/id_rsa)"));
-}, { code: 1, synthetic: true, includes: ["fora do projeto"] });
-
-test("G11: comando fixo destrutivo e apanhado", (dir) => {
-  patchSettings(dir, (c) => c.permissions.allow.push("Bash(rm -rf /tmp/x)"));
-}, { code: 1, synthetic: true, includes: ["shell interactiva ou comando destrutivo"] });
-
-test("G11: allow que cai dentro de um prefixo em ask e contradicao", (dir) => {
-  patchSettings(dir, (c) => {
-    c.permissions.ask = ["Bash(git push:*)"];
-    c.permissions.allow.push("Bash(git push origin:*)");
-  });
-}, { code: 1, synthetic: true, includes: ["cai dentro de", "contradicao"] });
-
-test("CHECKS: versao de dependencia desatualizada na doc avisa", (dir) => {
-  writeF(dir, "package.json", JSON.stringify({ name: "x", version: "1.0.0", dependencies: { next: "16.2.2" } }));
-  writeF(dir, "src/docs/CHANGELOG.md", "# CL\n\n## [v1.0.0] - Atual\n");
-  writeF(dir, ".agent/rules/core-rules.md", "# core\n\nStack: Next.js 15.0.0\n");
-  const g = readF(dir, GUARD).replace(
-    "const CHECKS = [\n",
-    'const CHECKS = [\n  { name: "Next.js", pkg: "next", pattern: /Next\\.js\\s+(\\d+(?:\\.\\d+(?:\\.\\d+)?)?)/g, files: [".agent/rules/core-rules.md"] },\n'
-  );
-  writeF(dir, GUARD, g);
-}, { code: 1, synthetic: true, includes: ["documentado 15.0.0", "atual 16.2.2"] });
+registarSettings();
 
 // --- Ficheiros em branco: "existe mas vazio" != "ausente" ---------------------
 test("blank: AGENTS.md vazio nao passa a verde", (dir) => {
@@ -827,103 +392,7 @@ test("G8: conta so os imports que RESOLVEM", (dir) => {
   includes: ["2 de 3 @imports"],
 });
 
-// --- Guard 12: a contagem da checklist como dado derivado ---------------------
-// Quatro sitios de aviso, um teste por sitio. O guard nasceu porque a checklist cresceu
-// para 25 e as quatro citacoes em prosa ficaram em 24 sem nada notar.
-const SYNC = ".agent/rules/sync-docs.md";
 
-test("G12a: numeracao com salto avisa", (dir) => {
-  // 25 pontos escritos 1..24 e depois "26." — o erro tipico ao inserir um ponto no meio.
-  writeF(dir, SYNC, readF(dir, SYNC).replace(/^25\. \[ \]/m, "26. [ ]"));
-}, { code: 1, includes: ["numeracao da checklist quebrada", 'ponto #25 esta escrito como "26."'] });
+registarDerivedCounts();
 
-test("G12a: numero duplicado avisa", (dir) => {
-  writeF(dir, SYNC, readF(dir, SYNC).replace(/^5\. \[ \]/m, "4. [ ]"));
-}, { code: 1, includes: ["numeracao da checklist quebrada"] });
-
-test("G12a: checklist sem pontos avisa", (dir) => {
-  // Formato mudado: os pontos passam a bullets. A contagem deixa de existir.
-  writeF(dir, SYNC, readF(dir, SYNC).replace(/^\d+\. \[ \]/gm, "- [ ]"));
-}, { code: 1, includes: ["nao tem pontos de checklist"] });
-
-test("G12b: citacao desatualizada avisa (o defeito original)", (dir) => {
-  // Exatamente o estado em que o repo estava: checklist a 25, prosa a 24.
-  writeF(dir, ".agent/rules/process-rules.md",
-    readF(dir, ".agent/rules/process-rules.md").replace("(25 pontos)", "(24 pontos)"));
-}, { code: 1, includes: ['diz "(24 pontos" mas a checklist de sync-docs.md tem 25'] });
-
-test("G12b: ponto novo sem atualizar a prosa avisa", (dir) => {
-  // O inverso: cresce a checklist e ninguem toca nas citacoes.
-  writeF(dir, SYNC, readF(dir, SYNC) + "\n26. [ ] ponto novo\n");
-}, { code: 1, includes: ['diz "(25 pontos" mas a checklist de sync-docs.md tem 26'] });
-
-test("G12b: '(N pontos' de OUTRA contagem nao e falso positivo", (dir) => {
-  // O relatorio de fecho de sprint tem 6 pontos e e citado com a mesma forma. Uma versao
-  // anterior deste guard reclamava dele. A linha nao menciona sync-docs, logo nao conta.
-  writeF(dir, "src/docs/agent-guide.md",
-    readF(dir, "src/docs/agent-guide.md") + "\n> relatorio de fecho (6 pontos — ver process-rules.md)\n");
-}, { code: 0, includes: ['citacao(oes) de "(N pontos" coerentes com 25'] });
-
-test("G12b: zero citacoes avisa (o ponteiro obrigatorio desapareceu)", (dir) => {
-  for (const f of [".agent/rules/process-rules.md", ".agent/workflows/review.md",
-                   ".agent/workflows/deploy.md", "src/docs/agent-guide.md"]) {
-    writeF(dir, f, readF(dir, f).replace(/\(25 pontos/g, "(a checklist"));
-  }
-}, { code: 1, includes: ["nenhum ficheiro cita o tamanho da checklist"] });
-
-test("G12: sync-docs.md ausente da SKIP visivel, nao silencio", (dir) => {
-  rmSync(file(dir, SYNC));
-}, { code: 0, includes: ["SKIP  Guard 12"] });
-
-// --- Guard 12c: contagem de fases do metodo por ticket ------------------------
-// O defeito original: o ficheiro tinha 6 fases (0-5) e tres sitios diziam "5 fases",
-// omitindo sempre a Fase 4 (leitor independente) — a mais cara. Num ficheiro que afirma
-// "nenhuma se salta", faltar uma nos resumos e um erro de conteudo, nao uma gralha.
-const METODO = ".agent/rules/ticket-method.md";
-
-test("G12c: citacao em digito desatualizada avisa (o defeito original)", (dir) => {
-  writeF(dir, ".agent/rules/process-rules.md",
-    readF(dir, ".agent/rules/process-rules.md").replace("passa por 6 fases", "passa por 5 fases"));
-}, { code: 1, includes: ['diz "5 fases" mas .agent/rules/ticket-method.md tem 6'] });
-
-test("G12c: citacao por palavra desatualizada avisa", (dir) => {
-  writeF(dir, METODO, readF(dir, METODO).replace("Seis fases:", "Cinco fases:"));
-}, { code: 1, includes: ['diz "Cinco fases" mas o metodo tem 6 (seis)'] });
-
-test("G12c: fase acrescentada sem atualizar a prosa avisa", (dir) => {
-  writeF(dir, METODO, readF(dir, METODO) + "\n## Fase 6 — Algo novo\n\nconteudo.\n");
-}, { code: 1, includes: ['fases" mas'] });
-
-test("G12c: fase renumerada com salto avisa", (dir) => {
-  writeF(dir, METODO, readF(dir, METODO).replace("## Fase 4 —", "## Fase 7 —"));
-}, { code: 1, includes: ["esperado 0..5 sem saltos"] });
-
-test("G12c: metodo sem cabecalhos `## Fase N` avisa", (dir) => {
-  writeF(dir, METODO, readF(dir, METODO).replace(/^## Fase \d+/gm, "## Etapa"));
-}, { code: 1, includes: ["nao tem cabecalhos `## Fase N`"] });
-
-test("G12c: zero citacoes avisa (o ponteiro obrigatorio desapareceu)", (dir) => {
-  for (const f of [".agent/rules/process-rules.md", METODO, "src/docs/agent-guide.md"]) {
-    writeF(dir, f, readF(dir, f).replace(/(\d+|Seis|seis)\s+fases/g, "as etapas"));
-  }
-}, { code: 1, includes: ["nenhum ficheiro cita o numero de fases"] });
-
-test("G12c: metodo ausente da SKIP visivel, nao silencio", (dir) => {
-  rmSync(file(dir, METODO));
-}, { code: 0, includes: ["SKIP  Guard 12c"] });
-
-// --- Resumo ------------------------------------------------------------------
-console.log("");
-console.log(`  ${passed} passaram, ${failures.length} falharam.`);
-if (failures.length) {
-  console.log("\n--- Detalhe das falhas ---");
-  for (const f of failures) {
-    console.log(`\n[${f.name}]`);
-    for (const p of f.problems) console.log(`  ${p}`);
-    console.log(f.out.split("\n").map((l) => `    | ${l}`).join("\n"));
-  }
-  console.log("");
-  process.exit(1);
-}
-console.log("\nTodos os testes dos guards passaram.\n");
-process.exit(0);
+resumo();
