+++ README.md
@@ -81,9 +81,16 @@ When you open a new AI session in any project using this template, the agent **a
 │   └── market-scan.md          <- /market-scan — Market/competitor analysis + feature ideation
 └── scripts/
     ├── check-bundle-sizes.mjs  <- Bundle size checker (Next.js)
-    ├── check-doc-versions.mjs  <- Doc guards: rules byte-budget, CLAUDE/GEMINI parity, CHANGELOG, versions
+    ├── check-doc-versions.mjs  <- Doc guards: entry point + doc-parity guards
+    ├── guards/                 <- Guard modules split out of the entry point
+    │   ├── settings.mjs        <- Guard 11: .claude/settings.json permission boundary
+    │   ├── versions.mjs        <- Guard 3 + documented dependency versions
+    │   └── derived-counts.mjs  <- Guards 12/12c: counts cited in prose, recomputed
     ├── check-backlog.mjs       <- Backlog counters/progress + duplicate-ID checker
-    ├── test-guards.mjs         <- Negative tests for the doc guards (no deps, no package.json)
+    ├── test-guards.mjs         <- Entry point for the doc-guard suites (no deps, no package.json)
+    ├── test-harness.mjs        <- Shared sandbox + test() + summary
+    ├── tests-settings.mjs      <- Guard 11 tests (mirrors guards/settings.mjs)
+    ├── tests-derived-counts.mjs<- Guards 12/12c tests (mirrors guards/derived-counts.mjs)
     ├── test-bundle-sizes.mjs   <- Negative tests for the bundle checker (no Next.js needed)
     ├── test-backlog.mjs        <- Negative tests for the backlog checker (synthetic fixture)
     ├── mutation-sweep.mjs      <- Proves the suites assert: disables each warning, demands red
